+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.7.2] - 28-01-26
+
 ### Fixed
 * Remove `tnt_crud_storage_safe_mode_enabled` metric from router.